feat(safe): Add Safe object to dom.
Simplify packaging + .node loading + testcafe running
